Work around a hw hang if halting a wave at s_endpgm
In current gfxips, halting a wave at the s_endpgm instruction causes the chip
to hang. To avoid it, for each trap, we read the instruction at the pc, and if
it is s_endpgm, we skip setting the halt bit. The wave will continue and
the debugger will detect that is has terminated.

Change-Id: I71c83d1d1ed779bca016031510daa01a7db50b9a