Add GpuID to QueueInfo and CodeObjectInfo
NodeIDs are not portable across processes, the Debugger API should use GpuIds
to identify agents.

Improve GetGpuId to find a node with a matching LocationId instead of looking
at the NodeId's entry in the sysfs. NodeId and SysfsId may be different if
using cgroups, or environment variables that affect the agent enumeration.

Change-Id: Iea4a46d47372042aa50476d6f178622fd186b61d