Add functions for gdb to set breakpoints on
Add _loader_debug_state after each modification to the code object
list. This is temporary, until we can get the list directly from
the ROCR.

Add OnLoadComplete at the end of the debug agent initialization. This
is also temporary, until we are able to create a kfd_process on
behalf of another process. We also need the debug agent to allocate
the stepping buffer, and install the trap handler.

Change-Id: Ia40017ac3138caa12cd2a5b66c36120f3f74142f