Build rocr_debug_agent_test - NDEBUG
SWDEV-207576

There is also code to add the libhsakmt.so.1 library explicitly into
the link path. This is to work around a bug [1] in the compile time
linker. The runtime linker recognises a number of special strings
"$ORIGIN", "${ORIGIN}", "$LIB" etc in the DT_RUNPATH and DT_RPATH.
Prior to version 2.28 of /bin/ld, the compile time linker didn't
support these, and so looked for directories with the literal name
$ORIGIN rather than substituting the location of the ELF object with
the specified DT_RUNPATH. Ubuntu 16.04 has version 2.26.1.

[1] https://sourceware.org/bugzilla/show_bug.cgi?format=multiple&id=20535

Change-Id: I3f34d157c9dcc0f996ccadc9c4c7daf53ddae92c
Signed-off-by: Icarus Sparry <icarus.sparry@amd.com>