Improve process tear-down
Given how runtime works, it is possible that
debug_agent_hsa_executable_destroy gets called after the debug agent's
global destructor are called.  In such situation,
debug_agent_hsa_executable_destroy ends up using ghost objects, which is
undefined behavior and can lead to errors.

This patch does a best effort at detecting such situation by introducing
a bool flag which changes when we are about to reset the objects.  Seeing
that this flag has changed is a guarantee that the objects cannot be
used anymore, but unfortunately the opposite is not true.

Change-Id: I67a94f93941086a846b07f5c779bf31455a63805