Do not use static storage duration for the worker thread
Rocr debug agent works by spawning a worker thread which uses dbgapi to
monitor the rocr activity of the current process.  The current
implementation uses a std::optional<DebugAgentWorker> with static
storage duration to keep track of this thread.  This approach means that
on process exit, the optional's dtor is called.  One problem which
showed in the past is that the runtime can still, after globals of the
debug agent have been destructed, call into its callback.  This leads to
use-after-delete bugs.

A first approach was to use an atomic flag to indicate that the globals
have been destructed and became unsafe to use.  This approach could
still fail in multi-threaded environments, where dtors can be called
while objects are in use.

After discussion with others, this patch follows another approch which
is to not use global with static storage lifetime.  Instead, this patch
statically allocates memory for the object, and constructs an instance
in place.  This way, the destructor is never called, making the
std::optional<DebugAgentWorker> always safe to use.

This object is wrapped with a mutex to synchronize access to the object.
Access to the object can only be done via get_worker_thread.  This
function returns a WorkerThreadAccess which owns a lock on the mutex and
can forward requests if the worker thread is active.

Change-Id: Icb099a0f77ed8e612ab702ec511a606f9fb7f419