test/run-test.py: Ignore "architecture not supported" warning
Some systems can have multiple GPUs, with some of them not being
supported by dbgapi.  Debug agent is still functional on the remaining
GPUs.

In such configuration, we have 2 options:
- Either the compute workload is only executed on the supported devices,
  in which case the warning should not cause an error (we can ignore
it),
- or the workload is executed on the unsupported GPU and even if we
  ignore the warning the tests will eventually fail.

In any case, it is safe to ignore the "architecture not supported"
warning in run-test.py.  This patch adds it to the `filter_warning`
predicate to avoid spurious failure.

Change-Id: I084b6bc68dbb361f81b972a524d58b19ffb96709