Properly use /proc/PID/mem's open FD as client_process_id
The client_process_id given to rocm-dbgapi on process_attach is not
matching what the amd_dbgapi_xfer_global_memory callback expects.  It
gets given a pointer to the dbgapi_process_id instead of the expected
pointer to the opened file descriptor to /proc/PID/mem.

This patch fixes this oversight.

Change-Id: Iade686b6ae5d65b8d52ec66837e8f88a1ede7cff